requirements.txt
django==3.2
environs==11.0.0
django-filter==24.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.32
pillow==10.2.0
